requirements.txt
torch==2.0.0
torchvision==0.15.0
Pillow==9.5.0
streamlit==1.19.0
matplotlib==3.7.1
tensorboard==2.12.0